# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "**TMF API Reference : TMF - 724 Incident**\n\n**Release : 21.5 - June\
    \ 2021**\n\nThe Incident API provides a standardized client interface to Incident\
    \ Management Systems for creating, tracking and managing Incidents as a result\
    \ of an issue or problem identified by a customer or another system. \nExamples\
    \ of Incident API originators (clients) include CRM applications, network management\
    \ or fault management systems, or other Incident management systems (e.g. B2B).\n\
    \nThe API supports the ability to send requests to create a new Incident specifying\
    \ the nature and severity of the trouble or issue as well as all necessary related\
    \ information. The API also includes mechanisms to search for and update existing\
    \ Incidents. Notifications are defined to provide information when a Incident\
    \ has been updated, including status changes. A basic set of states of a Incident\
    \ has been specified (as an example) to handle Incident lifecycle management.\n\
    Incidenting API manages Incident resource:\n\n  -\tA Incident represents a record,\
    \ or an issue raised by requestor that need to be solved, used for reporting and\
    \ managing the resolution of problems, incidents or request -\tMain Incident attributes\
    \ are its description, severity, type, related dates (creation, expected resolution,\
    \ resolution), state and related information (change reason and change date),\
...
host: "serverRoot:8080"
basePath: "/tmf-api/Incident/v4/"
...
schemes:
...
consumes:
...
produces:
...
paths:
...
  /incidentTask:
    get:
      tags:
      - "incidentTask"
      summary: "List or find IncidentTask objects"
      description: "This operation list or find IncidentTask entities"
      operationId: "listIncidentTask"
      parameters:
      - name: "fields"
        in: "query"
        description: "Comma-separated properties to be provided in response"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "Requested index for start of resources to be provided in response"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "Requested number of resources to be provided in response"
        required: false
        type: "integer"
      responses:
        200:
          description: "Success"
          headers:
            X-Result-Count:
              type: "integer"
              description: "Actual number of items returned in the response body"
            X-Total-Count:
              type: "integer"
              description: "Total number of items matching criteria"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/IncidentTask"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "IncidentTask"
    post:
      tags:
      - "incidentTask"
      summary: "Creates a IncidentTask"
      description: "This operation creates a IncidentTask entity."
      operationId: "createIncidentTask"
      parameters:
      - in: "body"
        name: "incidentTask"
        description: "The IncidentTask to be created"
        required: true
        schema:
          $ref: "#/definitions/IncidentTask_Create"
      responses:
        201:
          description: "Created"
          schema:
            $ref: "#/definitions/IncidentTask"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "IncidentTask"
  /incidentTask/{id}:
    get:
      tags:
      - "incidentTask"
      summary: "Retrieves a IncidentTask by ID"
      description: "This operation retrieves a IncidentTask entity. Attribute selection\
        \ is enabled for all first level attributes."
      operationId: "retrieveIncidentTask"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the IncidentTask"
        required: true
        type: "string"
      - name: "fields"
        in: "query"
        description: "Comma-separated properties to provide in response"
        required: false
        type: "string"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/IncidentTask"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "IncidentTask"
    delete:
      tags:
      - "incidentTask"
      summary: "Deletes a IncidentTask"
      description: "This operation deletes a IncidentTask entity."
      operationId: "deleteIncidentTask"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the IncidentTask"
        required: true
        type: "string"
      responses:
        204:
          description: "Deleted"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "IncidentTask"
    patch:
      tags:
      - "incidentTask"
      summary: "Updates partially a IncidentTask"
      description: "This operation updates partially a IncidentTask entity."
      operationId: "patchIncidentTask"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the IncidentTask"
        required: true
        type: "string"
      - in: "body"
        name: "incidentTask"
        description: "The IncidentTask to be updated"
        required: true
        schema:
          $ref: "#/definitions/IncidentTask_Update"
      responses:
        200:
          description: "Updated"
          schema:
            $ref: "#/definitions/IncidentTask"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "IncidentTask"
...
definitions:
...
  IncidentTask:
    type: "object"
    properties:
      id:
        type: "string"
        description: "unique identifier"
      href:
        type: "string"
        format: "uri"
        description: "Hyperlink reference"
      eventId:
        type: "integer"
        description: "The ID of the incident id"
      name:
        type: "string"
        description: "The name of Task, for example:RAN incident diagnose"
      status:
        type: "string"
        description: "The status of the task , for example:running, finished,canceled\
          \ etc"
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class"
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class Extensible name"
    description: "Incident task on the incident, for example : diagnose, repair etc"
    example:
      eventId: 0
      '@baseType': "@baseType"
      '@type': "@type"
      name: "name"
      id: "id"
      href: "http://example.com/aeiou"
      '@schemaLocation': "http://example.com/aeiou"
      status: "status"
  IncidentTask_Create:
    type: "object"
    required:
    - "eventId"
    - "name"
    properties:
      eventId:
        type: "integer"
        description: "The ID of the incident id"
      name:
        type: "string"
        description: "The name of Task, for example:RAN incident diagnose"
      status:
        type: "string"
        description: "The status of the task , for example:running, finished,canceled\
          \ etc"
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class"
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class Extensible name"
    description: "Incident task on the incident, for example : diagnose, repair etc\n\
      Skipped properties: id,href"
    example:
      eventId: 0
      '@baseType': "@baseType"
      '@type': "@type"
      name: "name"
      '@schemaLocation': "http://example.com/aeiou"
      status: "status"
  IncidentTask_Update:
    type: "object"
    properties:
      name:
        type: "string"
        description: "The name of Task, for example:RAN incident diagnose"
      status:
        type: "string"
        description: "The status of the task , for example:running, finished,canceled\
          \ etc"
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class"
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class Extensible name"
    description: "Incident task on the incident, for example : diagnose, repair etc\n\
      Skipped properties: id,href,eventId"
    example:
      '@baseType': "@baseType"
      '@type': "@type"
      name: "name"
      '@schemaLocation': "http://example.com/aeiou"
      status: "status"
...
  Error:
    type: "object"
    required:
    - "code"
    - "reason"
    properties:
      code:
        type: "string"
        description: "Application relevant detail, defined in the API or a common\
          \ list."
      reason:
        type: "string"
        description: "Explanation of the reason for the error which can be shown to\
          \ a client user."
      message:
        type: "string"
        description: "More details and corrective actions related to the error which\
          \ can be shown to a client user."
      status:
        type: "string"
        description: "HTTP Error code extension"
      referenceError:
        type: "string"
        format: "uri"
        description: "URI of documentation describing the error."
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class."
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name."
    description: "Used when an API throws an Error, typically with a HTTP error response-code\
      \ (3xx, 4xx, 5xx)"